+++ app/channels/map_channel.rb
@@ -44,6 +44,6 @@ class MapChannel < ApplicationCable::Channel
   end
 
   def map_atts(data)
-    data.slice("name", "description", "base_map", "center", "zoom", "pitch", "terrain")
+    data.slice("name", "description", "base_map", "center", "zoom", "pitch", "bearing", "terrain")
   end
 end