+++ app/javascript/maplibre/styles.js
@@ -3,6 +3,7 @@ import * as f from 'helpers/functions'
 import { showFeatureDetails } from 'maplibre/modals'
 
 let highlightedFeatureId
+let stickyFeatureHighlight = false
 export const viewStyleNames = [
   'polygon-layer',
   'polygon-layer-extrusion',
@@ -17,46 +18,56 @@ export const viewStyleNames = [
 
 export function resetHighlightedFeature (source = 'geojson-source') {
   if (highlightedFeatureId) {
-    map.setFeatureState({ source, id: highlightedFeatureId },
-      { active: false })
+    map.setFeatureState({ source, id: highlightedFeatureId }, { active: false })
     highlightedFeatureId = null
+    // drop feature param from url
+    const url = new URL(window.location.href)
+    url.searchParams.delete('f')
+    window.history.replaceState({}, document.title, url.toString())
   }
   // reset active modals
-  f.e('.map-modal', e => { e.style.display = 'none' })
+  f.e('#feature-details-modal', e => { e.classList.remove('show') })
 }
 
-export function highlightFeature (feature, source = 'geojson-source') {
+export function highlightFeature (feature, sticky = false, source = 'geojson-source') {
   resetHighlightedFeature()
   if (feature.id) {
+    stickyFeatureHighlight = sticky
     highlightedFeatureId = feature.id
     // load feature from source, the style only returns the dimensions on screen
     const sourceFeature = geojsonData.features.find(f => f.id === feature.id)
     showFeatureDetails(sourceFeature)
-    // A feature's state is not part of the GeoJSON or vector tile data
-    map.setFeatureState({ source, id: feature.id },
-      { active: true })
+    // A feature's state is not part of the GeoJSON or vector tile data but can get used in styles
+    map.setFeatureState({ source, id: feature.id }, { active: true })
+    // set url to feature
+    if (sticky) {
+      const newPath = `${window.location.pathname}?f=${feature.id}`
+      window.history.pushState({}, '', newPath)
+    }
   }
 }
 
 export function initializeViewStyles () {
   viewStyleNames.forEach(styleName => {
     map.addLayer(styles[styleName])
-
-    // click is only needed for mobile now
+    // click is needed to select on mobile and for sticky highlight
     map.on('click', styleName, function (e) {
       if (!e.features?.length || window.gon.map_mode === 'static') { return }
-      highlightFeature(e.features[0])
+      highlightFeature(e.features[0], true)
     })
   })
 
+  // highlight features on hover
   map.on('mousemove', (e) => {
-    resetHighlightedFeature()
-    const features = map.queryRenderedFeatures(e.point)
-    if (!features?.length || window.gon.map_mode === 'static') { return }
-    if (features[0].source === 'geojson-source') {
-      highlightFeature(features[0])
-    } else {
-      // console.log(features[0])
+    if (!(stickyFeatureHighlight && highlightedFeatureId)) {
+      resetHighlightedFeature()
+      const features = map.queryRenderedFeatures(e.point)
+      if (!features?.length || window.gon.map_mode === 'static') { return }
+      if (features[0].source === 'geojson-source') {
+        highlightFeature(features[0])
+      } else {
+        // console.log(features[0])
+      }
     }
   })
 
@@ -96,7 +107,7 @@ const fillOpacityActive = ['*', 0.7, fillOpacity]
 
 const lineColor = ['coalesce', ['get', 'stroke'], ['get', 'user_stroke'], featureColor]
 const lineWidth = ['to-number', ['coalesce',
-  ['get', 'stroke-width'], ['get', 'user_stroke-width'], 3]]
+  ['get', 'stroke-width'], ['get', 'user_stroke-width'], 2]]
 const lineWidthActive = ['+', 2, lineWidth]
 const lineOpacity = ['to-number', ['coalesce',
   ['get', 'stroke-opacity'], ['get', 'user_stroke-opacity'], 0.8]]
@@ -114,8 +125,8 @@ export const pointSize = ['to-number', ['coalesce',
   ['case',
     ['any', ['has', 'user_marker-symbol'], ['has', 'marker-symbol']],
     16, 6]]]
-const pointSizeActive = ['+', 2, pointSize]
-const pointOutlineSize = ['to-number', ['coalesce', ['get', 'user_stroke-width'], ['get', 'stroke-width'], 2]]
+export const pointSizeActive = ['+', 1, pointSize]
+export const pointOutlineSize = ['to-number', ['coalesce', ['get', 'user_stroke-width'], ['get', 'stroke-width'], 2]]
 export const pointOutlineSizeActive = ['+', 1, pointOutlineSize]
 const pointOutlineColor = ['coalesce', ['get', 'user_stroke'], ['get', 'stroke'], featureOutlineColor]
 const pointOpacity = 0.7
@@ -127,6 +138,13 @@ const pointOpacityActive = 0.9
 const iconSizeFactor = ['/', pointSize, 6]
 const iconSize = ['*', 1 / 8, iconSizeFactor]
 // const iconSizeActive = ['*', 1.1, iconSize] // icon-size is not a paint property
+const labelSize = ['to-number', ['coalesce', ['get', 'user_label-size'], ['get', 'label-size'], 16]]
+
+// font must be available via glyphs:
+// openmaptiles: https://github.com/openmaptiles/fonts/tree/gh-pages
+// maptiler: https://docs.maptiler.com/gl-style-specification/glyphs/
+// Emojis are not in the character range: https://github.com/maplibre/maplibre-gl-js/issues/2307
+const labelFont = ['Klokantech Noto Sans Bold']
 
 export const styles = {
   'polygon-layer': {
@@ -234,7 +252,9 @@ export const styles = {
     source: 'geojson-source',
     filter: ['all',
       ['==', '$type', 'Point'],
-      ['!=', 'active', 'true']
+      ['!=', 'active', 'true'],
+      ['!=', 'meta', 'midpoint'],
+      ['!=', 'meta', 'vertex']
     ],
     paint: {
       'circle-pitch-scale': 'map', // points get bigger when camera is closer
@@ -255,12 +275,7 @@ export const styles = {
           pointOpacity
         ]],
       'circle-blur': 0.1,
-      'circle-stroke-color': [
-        'case',
-        ['boolean', ['feature-state', 'active'], false],
-        featureColor,
-        pointOutlineColor
-      ],
+      'circle-stroke-color': pointOutlineColor,
       'circle-stroke-width': [
         'case',
         ['boolean', ['feature-state', 'active'], false],
@@ -314,12 +329,8 @@ export const styles = {
     filter: ['has', 'label'],
     layout: {
       'text-field': ['coalesce', ['get', 'label'], ['get', 'room']],
-      'text-size': 16,
-      // must be available via glyphs:
-      // openmaptiles: https://github.com/openmaptiles/fonts/tree/gh-pages
-      // maptiler: https://docs.maptiler.com/gl-style-specification/glyphs/
-      // Emojis are not in the character range: https://github.com/maplibre/maplibre-gl-js/issues/2307
-      'text-font': ['Klokantech Noto Sans Regular'],
+      'text-size': labelSize,
+      'text-font': labelFont,
       // arrange text to avoid collision
       'text-variable-anchor': ['top', 'bottom', 'left', 'right'],
       // distance the text from the element depending on the type
@@ -333,8 +344,8 @@ export const styles = {
       'text-ignore-placement': false // hide on collision
     },
     paint: {
-      'text-color': ['coalesce', ['get', 'label-color'], '#000'],
-      'text-halo-color': ['coalesce', ['get', 'label-shadow'], '#fff'],
+      'text-color': ['coalesce', ['get', 'user_label-color'], ['get', 'label-color'], '#000'],
+      'text-halo-color': ['coalesce', ['get', 'user_label-shadow'], ['get', 'label-shadow'], '#fff'],
       'text-halo-width': 1
     }
   }