+++ spec/factories/maps.rb
@@ -1,6 +1,6 @@
 FactoryBot.define do
   factory :map do
-    base_map { 'osmRasterTiles' }
+    base_map { 'test' }
     center { Map::DEFAULT_CENTER }
   end
 end