+++ spec/features/map_settings_spec.rb
@@ -20,7 +20,7 @@ describe 'Map' do
       find('.maplibregl-ctrl-map').click
       expect(page).to have_text('Configure Map')
       find(".layer-preview[data-base-map='stamenTonerTiles']").click
-      sleep(0.5) # make sure actioncable request is processed
+      expect(page).to have_text('Map properties updated')
       expect(map.reload.base_map).to eq 'stamenTonerTiles'
     end
 
@@ -28,57 +28,72 @@ describe 'Map' do
       find('.maplibregl-ctrl-map').click
       expect(page).to have_text('Configure Map')
       find('#map-terrain').click
-      sleep(0.5) # make sure actioncable request is processed
+      expect(page).to have_text('Map properties updated')
       expect(map.reload.terrain).to eq true
     end
   end
 
   context 'when map settings change server side' do
-    it 'basemap update' do
-      map.update(base_map: 'stamenTonerTiles')
-      sleep(0.5) # make sure actioncable request is processed
+    it 'name update' do
       find('.maplibregl-ctrl-map').click
+      map.update(name: 'New World')
+      expect(page).to have_text('New World')
+    end
+
+    it 'basemap update' do
+      find('.maplibregl-ctrl-map').click
+      map.update(base_map: 'stamenTonerTiles')
+      expect(page).to have_text('Map style loaded')
       expect(page).to have_css('.layer-preview[data-base-map="stamenTonerTiles"].active')
     end
 
     it 'terrain update' do
-      map.update(terrain: true)
-      sleep(0.5) # make sure actioncable request is processed
       find('.maplibregl-ctrl-map').click
+      map.update(terrain: true)
+      expect(page).to have_text('Map style loaded')
       expect(find('#map-terrain')).to be_checked
     end
 
     it 'map center update' do
+      find('.maplibregl-ctrl-map').click
       map.update(center: [ 11, 49.5 ])
       expect(page).to have_text('Map view updated')
-      sleep(2)
+      expect(page).to have_text('Center: 11,49.5')
       expect(page.evaluate_script("[map.getCenter().lng, map.getCenter().lat].toString()")).to eq('11,49.5')
     end
 
     it 'client follows default center update if map did not move' do
-      feature = create(:feature, :point, layer: map.layer, coordinates: [ 11, 49.5 ])
-      sleep(4)
+      find('.maplibregl-ctrl-map').click
+      feature = create(:feature, :point, layer: map.layer, coordinates: [ 11.543, 49.123 ])
+      expect(page).to have_text('Map view updated')
       # new default center are the feature coordinates
-      expect(page.evaluate_script("[map.getCenter().lng, map.getCenter().lat].toString()"))
+      expect(page.evaluate_script("[map.getCenter().lng.toFixed(3), map.getCenter().lat.toFixed(3)].toString()"))
         .to eq(feature.coordinates.join(','))
+      expect(page).to have_text("Center: #{feature.coordinates.join(',')} (auto)")
     end
 
     it 'map zoom update' do
+      find('.maplibregl-ctrl-map').click
       map.update(zoom: 16)
       expect(page).to have_text('Map view updated')
-      sleep(2)
       expect(page.evaluate_script("map.getZoom()")).to eq(16)
+      expect(page).to have_text('Zoom: 16')
     end
 
     it 'map pitch update' do
+      find('.maplibregl-ctrl-map').click
       map.update(pitch: 33)
       expect(page).to have_text('Map view updated')
-      sleep(2)
       expect(page.evaluate_script("map.getPitch()")).to eq(33)
+      expect(page).to have_text('Pitch: 33')
     end
 
     it 'map orientation update' do
-      # TODO
+      find('.maplibregl-ctrl-map').click
+      map.update(bearing: 33)
+      expect(page).to have_text('Map view updated')
+      expect(page.evaluate_script("map.getBearing()")).to eq(33)
+      expect(page).to have_text('Bearing: 33')
     end
   end
 end