+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## 2024-09
+
+* Direct link to map features
+
 
 ## 2024-07
 
+* sync center/zoom/pitch/bearing to other clients
 * geojson import
 * auto center + zoom
 * image upload in µlogger API